+++ tests/integration/test_api_query.py
@@ -151,6 +151,7 @@ def test_from_dict(api_config):
     assert isinstance(api, Client)
 
 
+@pytest.mark.timeout(180)
 def test_simple_query(api_config):
     config = ClientConfig(stream=False, response="json")
 
@@ -166,6 +167,7 @@ def test_simple_query(api_config):
     assert len(json.loads(result)["object"]) > 0
 
 
+@pytest.mark.timeout(180)
 def test_token(api_config):
     api = Client(
         auth={"token": api_config.api_token},
@@ -178,6 +180,7 @@ def test_token(api_config):
     assert len(json.loads(result)["object"]) > 0
 
 
+@pytest.mark.timeout(180)
 def test_query_id(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -191,6 +194,7 @@ def test_query_id(api_config):
     assert isinstance(len(json.loads(result)["object"]), int)
 
 
+@pytest.mark.timeout(180)
 def test_query_yesterday_to_today(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -205,6 +209,7 @@ def test_query_yesterday_to_today(api_config):
     assert len(json.loads(result)["object"]) == 1
 
 
+@pytest.mark.timeout(180)
 def test_query_from_seven_days(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -217,6 +222,7 @@ def test_query_from_seven_days(api_config):
     assert len(json.loads(result)["object"]) == 1
 
 
+@pytest.mark.timeout(180)
 def test_query_from_fixed_dates(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -235,6 +241,7 @@ def test_query_from_fixed_dates(api_config):
     assert len(json.loads(result)["object"]) == 1
 
 
+@pytest.mark.timeout(180)
 def test_stream_query(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -248,6 +255,7 @@ def test_stream_query(api_config):
     assert len(result) == 1
 
 
+@pytest.mark.timeout(180)
 def test_stream_query_no_results_bounded_dates(api_config):
     api = Client(
         auth={"key": api_config.api_key, "secret": api_config.api_secret},
@@ -285,6 +293,7 @@ def test_stream_query_no_results_unbounded_dates(api_config):
         assert False, "DevoClientException raised: %s" % (error)
 
 
+@pytest.mark.timeout(180)
 def test_pragmas(api_config):
     """Test the api when the pragma comment.free is used"""
     api = Client(
@@ -300,6 +309,7 @@ def test_pragmas(api_config):
     assert len(json.loads(result)["object"]) == 1
 
 
+@pytest.mark.timeout(180)
 def test_pragmas_not_comment_free(api_config):
     """Test the api when the pragma comment.free is not used"""
     api = Client(
@@ -316,6 +326,7 @@ def test_pragmas_not_comment_free(api_config):
 
 
 @pytest.mark.skip(reason="This is an internal functionality, not intended for external use")
+@pytest.mark.timeout(180)
 def test_unsecure_http_query(api_config):
     """
     This test is intended for checking unsecure HTTP requests. Devo will
@@ -436,6 +447,7 @@ def test_msgpack_future_queries(api_config):
     )
 
 
+@pytest.mark.timeout(180)
 def test_query_with_ip_as_int(api_config):
     config = ClientConfig(stream=False, response="json")
 
@@ -454,6 +466,7 @@ def test_query_with_ip_as_int(api_config):
     assert isinstance(resp_data[api_config.field_with_ip], int)
 
 
+@pytest.mark.timeout(180)
 def test_query_with_ip_as_string(api_config):
     config = ClientConfig(stream=False, response="json")